Refactor Star
